Use int instead of uint32 for all non-wire data

+++ binary/slice.go
@@ -11,7 +11,7 @@ import (
 // MakeSliceBuffer returns the slice type associated with c and t and with the given element count.
 // If the buffer is an slice which type matches with the expected by the acr then it will
 // be used as backing slice.
-func MakeSliceBuffer(c gltf.ComponentType, t gltf.AccessorType, count uint32, buffer any) any {
+func MakeSliceBuffer(c gltf.ComponentType, t gltf.AccessorType, count int, buffer any) any {
 	if buffer == nil {
 		return MakeSlice(c, t, count)
 	}
@@ -32,7 +32,7 @@ func MakeSliceBuffer(c gltf.ComponentType, t gltf.AccessorType, count uint32, bu
 // MakeSlice returns the slice type associated with c and t and with the given element count.
 // For example, if c is gltf.ComponentFloat and t is gltf.AccessorVec3
 // then MakeSlice(c, t, 5) is equivalent to make([][3]float32, 5).
-func MakeSlice(c gltf.ComponentType, t gltf.AccessorType, count uint32) any {
+func MakeSlice(c gltf.ComponentType, t gltf.AccessorType, count int) any {
 	var tp reflect.Type
 	switch c {
 	case gltf.ComponentUbyte:
@@ -63,17 +63,17 @@ func MakeSlice(c gltf.ComponentType, t gltf.AccessorType, count uint32) any {
 	case gltf.AccessorMat4:
 		tp = reflect.ArrayOf(4, reflect.ArrayOf(4, tp))
 	}
-	return reflect.MakeSlice(reflect.SliceOf(tp), int(count), int(count)).Interface()
+	return reflect.MakeSlice(reflect.SliceOf(tp), count, count).Interface()
 }
 
 // Type returns the associated glTF type data.
 // It panics if data is not an slice.
-func Type(data any) (c gltf.ComponentType, t gltf.AccessorType, count uint32) {
+func Type(data any) (c gltf.ComponentType, t gltf.AccessorType, count int) {
 	v := reflect.ValueOf(data)
 	if v.Kind() != reflect.Slice {
 		panic(fmt.Sprintf("gltf: binary.Type expecting a slice but got %s", v.Kind()))
 	}
-	count = uint32(v.Len())
+	count = v.Len()
 	switch data.(type) {
 	case []int8:
 		c, t = gltf.ComponentByte, gltf.AccessorScalar
